requirements.txt
pandas>=2.0.0
numpy>=1.21.0
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
cryptography>=3.4.0
requests>=2.25.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.14.0
dash-bootstrap-components>=1.4.0
jupyter>=1.0.0
